+++ src/models/ListData.ts
@@ -1,0 +1,76 @@
+export interface ListData {
+  RootFolder: RootFolder;
+  AllowContentTypes: boolean;
+  BaseTemplate: number;
+  BaseType: number;
+  ContentTypesEnabled: boolean;
+  CrawlNonDefaultViews: boolean;
+  Created: string;
+  CurrentChangeToken: CurrentChangeToken;
+  DefaultContentApprovalWorkflowId: string;
+  DefaultItemOpenUseListSetting: boolean;
+  Description: string;
+  Direction: string;
+  DisableGridEditing: boolean;
+  DocumentTemplateUrl?: string;
+  DraftVersionVisibility: number;
+  EnableAttachments: boolean;
+  EnableFolderCreation: boolean;
+  EnableMinorVersions: boolean;
+  EnableModeration: boolean;
+  EnableRequestSignOff: boolean;
+  EnableVersioning: boolean;
+  EntityTypeName: string;
+  ExemptFromBlockDownloadOfNonViewableFiles: boolean;
+  FileSavePostProcessingEnabled: boolean;
+  ForceCheckout: boolean;
+  HasExternalDataSource: boolean;
+  Hidden: boolean;
+  Id: string;
+  ImagePath: ImagePath;
+  ImageUrl: string;
+  IrmEnabled: boolean;
+  IrmExpire: boolean;
+  IrmReject: boolean;
+  IsApplicationList: boolean;
+  IsCatalog: boolean;
+  IsPrivate: boolean;
+  ItemCount: number;
+  LastItemDeletedDate: string;
+  LastItemModifiedDate: string;
+  LastItemUserModifiedDate: string;
+  ListExperienceOptions: number;
+  ListItemEntityTypeFullName: string;
+  MajorVersionLimit: number;
+  MajorWithMinorVersionsLimit: number;
+  MultipleDataList: boolean;
+  NoCrawl: boolean;
+  ParentWebPath: ImagePath;
+  ParentWebUrl: string;
+  ParserDisabled: boolean;
+  ServerTemplateCanCreateFolders: boolean;
+  TemplateFeatureId: string;
+  Title: string;
+  Url: string;
+}
+
+export interface ImagePath {
+  DecodedUrl: string;
+}
+
+export interface CurrentChangeToken {
+  StringValue: string;
+}
+
+export interface RootFolder {
+  Exists: boolean;
+  IsWOPIEnabled: boolean;
+  ItemCount: number;
+  Name: string;
+  ProgID?: any;
+  ServerRelativeUrl: string;
+  TimeCreated: string;
+  TimeLastModified: string;
+  UniqueId: string;
+  WelcomePage: string;
+}